requirements.txt
fastapi==0.111.0
alembic==1.13.1
uvicorn==0.29.0
# psycopg2==2.8.5
psycopg2-binary==2.9.5
SQLAlchemy==2.0.30
gunicorn==22.0.0
opencv-contrib-python==4.10.0.84
paho-mqtt==1.6.1